server/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv>=1.0.0
openai>=1.2.0
anthropic
voyageai
pdfplumber